api.util.js added
