+++ barf/api_7.js
@@ -1,78 +1,6 @@
-const queryString = require('query-string')
 const BigNumber = require('bignumber.js')
+const { constructMetadata, createQuery, post, getBlocUrl } = require('./api.util')
 const ax = require('./axios-wrapper')
-
-/**
- * This function constructes metadata that can be used to control the history and index flags
- * @method{constructMetadata}
- * @param{Object} options flags for history and indexing
- * @param{String} contractName
- * @returns{()} metadata
- */
-function constructMetadata(options, contractName) {
-  const metadata = {}
-  if (options === {}) return metadata
-
-  // history flag (default: off)
-  if (options.enableHistory) {
-    metadata.history = contractName
-  }
-  if (options.hasOwnProperty('history')) {
-    const newContracts = options.history.filter(contract => contract !== contractName).join()
-    metadata.history = `${options.history},${newContracts}`
-  }
-
-  // index flag (default: on)
-  if (options.hasOwnProperty('enableIndex') && !options.enableIndex) {
-    metadata.noindex = contractName
-  }
-  if (options.hasOwnProperty('noindex')) {
-    const newContracts = options.noindex.filter(contract => contract !== contractName).join()
-    metadata.noindex = `${options.noindex},${newContracts}`
-  }
-
-  // TODO: construct the "nohistory" and "index" fields for metadata if needed
-  // The current implementation only constructs "history" and "noindex"
-
-  return metadata
-}
-
-function createQuery(options) {
-  const queryObject = Object.assign(
-    { resolve: !options.isAsync },
-    options.stateQuery,
-  )
-  const query = `?${queryString.stringify(queryObject)}`
-  return query
-}
-
-async function post(url, endpoint, _body, options) {
-  function createBody(_body, options) {
-    // array
-    if (Array.isArray(_body)) return _body
-    // object
-    const body = Object.assign({}, _body)
-    const configTxParams = (options.config) ? options.config.txParams : undefined
-    // in order of priority: 1:body, 2:options, 3:config, 4:default
-    body.txParams = Object.assign(
-      { gasLimit: 32100000000, gasPrice: 1 },
-      configTxParams,
-      options.txParams,
-      _body.txParams,
-    )
-    return body
-  }
-
-  const body = createBody(_body, options)
-
-  return ax.post(url, endpoint, body, options)
-}
-
-function getBlocUrl(options) {
-  const node = options.node || 0
-  const nodeUrls = options.config.nodes[node]
-  return nodeUrls.blocUrl
-}
 
 async function getUsers(args, options) {
   const url = getBlocUrl(options)
@@ -83,7 +11,8 @@ async function getUsers(args, options) {
 async function getUser(args, options) {
   const url = getBlocUrl(options)
   const username = encodeURIComponent(args.username)
-  const endpoint = ('/users/:username').replace(':username', username)
+  const endpoint = ('/users/:username')
+    .replace(':username', username)
   return ax.get(url, endpoint, options)
 }
 
@@ -91,7 +20,8 @@ async function createUser(args, options) {
   const url = getBlocUrl(options)
   const username = encodeURIComponent(args.username)
   const data = { password: args.password }
-  const endpoint = ('/users/:username').replace(':username', username)
+  const endpoint = ('/users/:username')
+    .replace(':username', username)
   return ax.postue(url, endpoint, data, options)
 }
 
@@ -128,7 +58,8 @@ async function createContract(user, contract, options) {
 async function blocResults(hashes, options) { // TODO untested code
   const url = getBlocUrl(options)
   const query = createQuery(options)
-  const endpoint = `/transactions/results?${query}`
+  const endpoint = ('/transactions/results')
+    .concat(query)
   return post(url, endpoint, hashes, options)
 }
